+++ it/controllers/business/ScottishPartnershipNameControllerISpec.scala
@@ -70,7 +70,7 @@ class ScottishPartnershipNameControllerISpec extends ControllerISpec {
       val res: WSResponse = await(buildClient("/scottish-partnership-name").post(Map(ScottishPartnershipNameForm.scottishPartnershipNameKey -> testCompanyName)))
 
       res.status mustBe SEE_OTHER
-      res.header(HeaderNames.LOCATION) mustBe Some(controllers.applicant.routes.PartnershipIdController.startPartnerJourney.url)
+      res.header(HeaderNames.LOCATION) mustBe Some(controllers.grs.routes.PartnerPartnershipIdController.startPartnerJourney.url)
     }
 
     "return BAD_REQUEST for missing partnership name" in new Setup {
